[refac] MapInteractor 분리
- 클러스터 담당 Interactor와, POI add, remove, fetch 하는 Interactor 분리
