+++ Project17-C-Map/InteractiveClusteringMap/Map/ClusterInteractor.swift
@@ -14,37 +14,22 @@ protocol ClusterBusinessLogic: class {
     
 }
 
-protocol DataBusinessLogic: class {
-    
-    func add(tileId: CLong, poi: POI)
-    func remove(coordinate: Coordinate)
-    func fetch(coordinate: Coordinate) -> POIInfo?
-    
-}
-
-final class MapInteractor: MapBusinessLogic {
+final class ClusterInteractor: ClusterBusinessLogic {
     
     private let presenter: ClusterPresentationLogic
     private let quadTreeClusteringService: ClusteringServicing
     private let treeDataStore: TreeDataStorable
     
     init(treeDataStore: TreeDataStorable, presenter: ClusterPresentationLogic) {
-        self.presenter = presenter
         self.treeDataStore = treeDataStore
+        self.presenter = presenter
         self.quadTreeClusteringService = QuadTreeClusteringService(treeDataStore: treeDataStore)
     }
     
     func fetch(boundingBoxes: [CLong: BoundingBox], zoomLevel: Double) {
         boundingBoxes.forEach { tileId, boundingBox in
-            self.clustering(
-                tileId: tileId,
-                boundingBox: boundingBox,
-                zoomLevel: zoomLevel)
+            self.clustering(tileId: tileId, boundingBox: boundingBox, zoomLevel: zoomLevel)
         }
-    }
-    
-    func fetch(coordinate: Coordinate) -> POIInfo? {
-        return treeDataStore.fetch(coordinate: coordinate)
     }
     
     func remove(tileIds: [CLong]) {
@@ -65,16 +50,6 @@ final class MapInteractor: MapBusinessLogic {
             
             self.presenter.clustersToMarkers(tileId: tileId, clusters: clusters)
         }
-    }
-    
-    func add(tileId: CLong, poi: POI) {
-        treeDataStore.add(poi: poi)
-        presenter.add(tileId: tileId, poi: poi)
-    }
-    
-    func remove(coordinate: Coordinate) {
-        treeDataStore.remove(coordinate: coordinate)
-        presenter.delete(coordinate: coordinate)
     }
     
 }
