+++ Project17-C-Map/InteractiveClusteringMap/Map/POIDataInteractor.swift
@@ -1,0 +1,42 @@
+//
+//  DataInteractor.swift
+//  InteractiveClusteringMap
+//
+//  Created by eunjeong lee on 2020/12/20.
+//
+
+import Foundation
+
+protocol POIDataBusinessLogic: class {
+    
+    func add(tileId: CLong, poi: POI)
+    func remove(coordinate: Coordinate)
+    func fetch(coordinate: Coordinate) -> POIInfo?
+    
+}
+
+final class POIDataInteractor: POIDataBusinessLogic {
+    
+    private let presenter: DataPresentationLogic
+    private let treeDataStore: TreeDataStorable
+    
+    init(treeDataStore: TreeDataStorable, presenter: DataPresentationLogic) {
+        self.treeDataStore = treeDataStore
+        self.presenter = presenter
+    }
+    
+    func fetch(coordinate: Coordinate) -> POIInfo? {
+        return treeDataStore.fetch(coordinate: coordinate)
+    }
+    
+    func add(tileId: CLong, poi: POI) {
+        treeDataStore.add(poi: poi)
+        presenter.add(tileId: tileId, poi: poi)
+    }
+    
+    func remove(coordinate: Coordinate) {
+        treeDataStore.remove(coordinate: coordinate)
+        presenter.delete(coordinate: coordinate)
+    }
+    
+}
